engine/requirements.txt
fastapi==0.115.0
uvicorn==0.34.0
pyswisseph==2.10.3.2
kerykeion==4.20.1
python-dotenv==1.0.1
